pallas: support ordered jagged carry under fori_loop
Build the same sublane-aligned jagged window under fori_loop as
emit_pipeline, including aligned grid bounds, two-sided masks, truthful DMA
offsets, and full-block carry stores.

Add a default-config jagged BMM that previously could not lower correctly
through fori_loop. Run the carry and unsafe-write matrices under both
streaming lowerings.

+++ test/test_pallas.py
@@ -5179,7 +5179,10 @@ class TestPallas(TestCase):
         H, D = 4, 16
         offsets = torch.tensor([0, 10], device=DEVICE, dtype=torch.int32)
         torch.manual_seed(0)
-        y = torch.randn(H, 10, D, device=DEVICE, dtype=torch.bfloat16)
+        # Keep this mask-placement test directly addressable. A bf16 load would
+        # require a rounded-down window, while the major-dim output store cannot
+        # carry its head rows and is intentionally rejected.
+        y = torch.randn(H, 10, D, device=DEVICE, dtype=torch.float32)
 
         code, out = code_and_output(
             opposite,
